requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2023.11.17
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.1
daphne==4.1.2
distlib==0.3.8
Django==5.0
django-celery-beat==2.7.0
django-ckeditor==6.7.0
django-cors-headers==4.4.0
django-csp==3.7
django-jet-reboot==1.3.7
django-js-asset==2.2.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.13.3
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
kombu==5.4.2
Pillow==10.1.0
platformdirs==4.2.0
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
razorpay==1.4.1
redis==5.1.1
requests==2.31.0
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
zope.interface==7.0.3
tensorflow 
opennsfw2
django-channels-jwt
channels_redis
